+++ ProfessorManagement/Models/Professor_Grade.cs
@@ -7,8 +7,8 @@ namespace ProfessorManagement.Models
         [Key]
         public int Id { get; set; }
 
-        public byte ProfessorId { get; set; }
-        public byte GradeId { get; set; }
+        public int ProfessorId { get; set; }
+        public int GradeId { get; set; }
 
         public Professor professor { get; set; }
         public Grade grade { get; set; }
